Fix the ARJ build.
The script consists of just

SECTIONS
{
 /DISCARD/ : { *(.comment) *(.note) *(.stab) *(.stabstr) }
}

and is passed to ld without --script. The BFD behavior is to somehow
combine that with the default script. For exmaple, there is a default
start address and PT_LOAD areas are magically aligned.

I don't think that is something we can support in any sane way in lld.