Fix bash-static build.
This is an annoying one.

Configure has

AC_FUNC_MKTIME

which expands to a test of mktime that fails with the freebsd
implementation. Given that, bash compiles a mktime.o file that defines
just mktime and uses localtime. That goes in a .a file that is before
libc.

The freebsd libc defines mktime in localtime.o, which also defines
localtime among other functions.

When lld sees an undefined reference to mktime from libc, it uses the
bash provided one and then tries to find a definition of localtime. It
is found on libc's localtime.o, but now we have a duplicated error.

The reason it works with bfd is that bash doesn't use mktime directly
and the undefined reference from libc is resolved to the libc
implementation. It would also fail to link if bash itself directly
used mktime.

Freebsd should probably consider fixing mktime to pass AC_FUNC_MKTIME
(or fix AC_FUNC_MKTIME). It should probably also split mktime into its
own file so that it is possible to replace just that function.