Apply reproducible builds timestamp to non-python only
Python ports may produce a .pyo or .pyc file that contains a timestamp
corresponding to the source file from which it was built.  A mismatch
causes the file to be rebuilt or remain unused.

We need a better fix for this issue (and we could still set the
timestamp for those ports do not include .pyo/.pyc files), but this
should allow us to make other (non-Python) ports reproducible.

Reported by:	antoine