security/openssl*: fix build with lld as system linker
OpenSSL's build infrastructure tries to determine the type of linker in
use by grepping for '^GNU ld'.  lld is compatible with GNU ld and even
includes 'GNU' in its version information for these sorts of tests,
which are usually a little less strict.

Add a REINPLACE_CMD to replace '^GNU ld' with just 'GNU' so it will
tread lld the same as GNU ld.

PR:		214864, 231459
Submitted by:	Nathan